// superset-frontend/src/pages/RoleAssignment/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
export type RoleAssignmentAction = 'add' | 'replace' | 'remove';

export type RoleAssignmentScope = 'selected' | 'filtered';

export type RoleOption = {
  id: number;
  name: string;
};

export type RoleSummary = {
  id: number;
  name: string;
};

export type GroupSummary = {
  id: number;
  name: string;
};

export type UserOption = {
  active: boolean;
  id: number;
  first_name: string;
  last_name: string;
  username: string;
  email: string;
  roles: RoleSummary[];
  groups: GroupSummary[];
};

export type RoleAssignmentRequest = {
  roleIds: number[];
  action: RoleAssignmentAction;
  scope: RoleAssignmentScope;
  userIds?: number[];
  filters?: RoleAssignmentFilter[];
};

export type RoleAssignmentFilter = {
  col: string;
  opr: string;
  value: string | number | boolean | string[] | number[];
};